+++ server/components/components.spec.js
@@ -1,18 +1,15 @@
 /* global describe, it */
 /**
- * Test if sandbox have all services that need the application components.
- * @module sandbox.spec
+ * Test components module
+ * @module components.spec
  */
 'use strict'
 
-// Set test environment
-process.env.NODE_ENV = 'test'
+// Require first the server
+const server = require('../server')
 
 // npm modules
 const chai = require('chai')
-
-// Server
-const server = require('../../server')
 
 // Constants & Variables
 const sandbox = server.sandbox
@@ -20,8 +17,8 @@ const sandbox = server.sandbox
 // Chai styles
 const expect = chai.expect
 
-// Run test for api sandbox
-describe('Sandbox content', () => {
+// Parent blck
+describe('components', () => {
   // Services from user component
   it('User should register services', () => {
     expect(sandbox.hasAnswer('findingUserById')).to.equal(true)