+++ server/components/user/user-error.map.js
@@ -1,51 +1,45 @@
 /**
- * Templates for custom error for responses.
- * @module res-error-templates
- * @memberOf response-error
+ * A map with custom error for user component
+ * @module user-error.map
  */
 'use strict'
 
-/**
- * List all avaiable errors
- * Error template object:
-    {
-      name: 'ErrorName',
-      messages: 'Shit happends',
-      status: 404,
-      field: 'email' // Useful to identified the error client forms
-    }
- */
-// =========================
-module.exports = Object.freeze({
+// Create custom errors
+const errors = new Map([
   /** Used for email that have bad syntax */
-  InvalidEmail: {
+  ['InvalidEmail', {
     name: 'InvalidEmail',
     message: 'Invalid email',
     status: 422,
     field: 'email'
-  },
+  }],
 
   /** Used when role is invalid */
-  InvalidRole: {
+  ['InvalidRole', {
     name: 'InvalidRole',
     message: 'Invalid role',
     status: 422,
     field: 'role'
-  },
+  }],
 
   /** The user already is in db */
-  EmailAlreadyExits: {
+  ['EmailAlreadyExits', {
     name: 'EmailAlreadyExits',
     message: 'Email is already taken',
     status: 422,
     field: 'email'
-  },
+  }],
 
   /** Could not find the user */
-  UserNotFound: {
+  ['UserNotFound', {
     name: 'UserNotFound',
     message: 'User not found',
     status: 404,
     field: 'email'
-  }
-})
+  }]
+])
+
+/**
+ * Map of errors for user component
+ */
+module.exports = errors