+++ server/components/user/user.schema.js
@@ -12,6 +12,11 @@ const database = require('../../database')
 
 // Constants & Variables
 const Schema = mongoose.Schema
+const schemaOptions = {
+  toObject: { virtuals: true },
+  toJSON: { virtuals: true },
+  autoIndex: false
+}
 
 // Get database connection from a pool
 const db = database.dbPool.connections.default
@@ -34,12 +39,7 @@ const UserSchema = new Schema({
     type: Date,
     default: Date.now
   }
-},
-  {
-    toObject: { virtuals: true },
-    toJSON: { virtuals: true },
-    autoIndex: false
-  }) // EoS
+}, schemaOptions) // EoS
 
 /** User mongoose model */
 module.exports = db.model('User', UserSchema)