+++ server/response-error/index.js
@@ -1,4 +1,0 @@
-/**
- * @module response-error
- */
-module.exports = require('./response-error.service')